practice1/requirements.txt
fastapi==0.103.0
uvicorn==0.23.0
numpy==1.24.0
matplotlib==3.6.0
scipy==1.10.0
pillow==9.2.0
